Prevents admin account deletion
Adds a check to prevent deletion of admin accounts
via the deleteMyAccount endpoint.

This enhances account security by restricting admin
account removal through the standard user deletion
process.

+++ api/src/Controllers/AccountController.ts
@@ -1,7 +1,8 @@
-import { JsonController, Param, Body, Get, Post, Put, Delete, HeaderParam, CurrentUser, Authorized, Patch, BadRequestError, HttpCode, OnUndefined } from 'routing-controllers';
+import { JsonController, Param, Body, Get, Post, Put, Delete, HeaderParam, CurrentUser, Authorized, Patch, BadRequestError, HttpCode, OnUndefined, ForbiddenError } from 'routing-controllers';
 import { FilledUser, User } from "../Models/UserModel"
 import { deleteMyAccount, getUserById, updateMyAccount } from '../Services/users/usersPublic';
 import { zUpdateAccount } from '../Validators/users';
+import { UserRole } from '../DB_Schema/UserSchema';
 
 @JsonController("/account")
 export class AccountController {
@@ -27,6 +28,11 @@ export class AccountController {
   @Authorized()
   @HttpCode(204)
   async deleteMyAccount(@CurrentUser() user: User): Promise<boolean>{
+
+    if(user.role == UserRole.admin){
+      throw new ForbiddenError("Vous ne pouvez pas supprimer votre compte, car vous êtes admin")
+    }
+
     if (!await deleteMyAccount(user)) {
       throw new BadRequestError("Erreur lors de la suppression du compte");
     }
